Extract backdrop helper to nwg-dock-common::layer_shell
Move the per-monitor fullscreen-backdrop construction from
nwg-notifications into a shared module so the drawer can use it next
without copy-paste. Also adds an `exclude_connector` parameter so
callers that already render their own full-screen surface on a
specific monitor (the drawer's use-case) can avoid an input-region
race between the backdrop and their own surface on that output.

Notifications panel and DND menu pass `None` for the exclusion since
they're popups that don't cover any monitor themselves. Behavior is
identical to before for both.
