+++ crates/nwg-dock-common/src/lib.rs
@@ -4,6 +4,7 @@ pub mod desktop;
 pub mod error;
 pub mod hyprland;
 pub mod launch;
+pub mod layer_shell;
 pub mod pinning;
 pub mod process;
 pub mod signals;
